pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "GHOST MODE"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.29.0",
    "fastapi[standard]>=0.136.0",
    "greenlet>=3.0.0",
    "psycopg[binary]>=3.3.3",
    "pytest>=9.0.3",
    "sqlalchemy>=2.0.49",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.9.0.post0",
    "google-generativeai>=0.8.6",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
